refactor(rfid): unify pending scan modal into shared component
- Create PendingScanModal component in src/components/
- Remove duplicate keyframes, useEffect, and modal JSX from both pages
- Both ActivityScanningPage and TagAssignmentPage now use <PendingScanModal />
- Component is self-contained: reads store state, handles keyframe injection
