+++ src/pages/TagAssignmentPage.tsx
@@ -4,6 +4,7 @@ import { useState, useEffect, useCallback, useRef } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
 
 import { BackgroundWrapper } from '../components/background-wrapper';
+import { PendingScanModal } from '../components/PendingScanModal';
 import { ErrorModal, ModalBase } from '../components/ui';
 import BackButton from '../components/ui/BackButton';
 import { api, type TagAssignmentCheck } from '../services/api';
@@ -15,29 +16,6 @@ import { logNavigation, logUserAction, logError, createLogger } from '../utils/l
 import { safeInvoke, isTauriContext, isRfidEnabled } from '../utils/tauriContext';
 
 const logger = createLogger('TagAssignmentPage');
-
-// CSS keyframes for pending scan pulse animation (injected into head)
-const PENDING_SCAN_KEYFRAMES = `
-@keyframes pulse-scale {
-  0%, 100% {
-    transform: scale(1);
-    opacity: 1;
-  }
-  50% {
-    transform: scale(1.15);
-    opacity: 0.8;
-  }
-}
-
-@keyframes spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-`;
 
 /**
  * Helper to get assigned person from TagAssignmentCheck
@@ -73,7 +51,6 @@ interface RfidScannerStatus {
  */
 function TagAssignmentPage() {
   const { authenticatedUser, showPendingScan, clearPendingScan } = useUserStore();
-  const { activePendingScan } = useUserStore(state => state.rfid);
   const navigate = useNavigate();
   const location = useLocation();
 
@@ -131,23 +108,6 @@ function TagAssignmentPage() {
       }
       if (pendingScanTimerRef.current) {
         clearTimeout(pendingScanTimerRef.current);
-      }
-    };
-  }, []);
-
-  // Inject keyframes for pending scan animation
-  useEffect(() => {
-    const styleId = 'pending-scan-keyframes';
-    if (!document.getElementById(styleId)) {
-      const styleElement = document.createElement('style');
-      styleElement.id = styleId;
-      styleElement.textContent = PENDING_SCAN_KEYFRAMES;
-      document.head.appendChild(styleElement);
-    }
-    return () => {
-      const existing = document.getElementById(styleId);
-      if (existing) {
-        existing.remove();
       }
     };
   }, []);
@@ -550,112 +510,8 @@ function TagAssignmentPage() {
             </button>
           </ModalBase>
 
-          {/* Pending Scan Modal - shows immediate "detected" feedback while API processes */}
-          {activePendingScan && (
-            <ModalBase
-              isOpen={!!activePendingScan}
-              onClose={() => {
-                /* No auto-close for pending modal - it transitions to result */
-              }}
-              backgroundColor="#3B82F6"
-            >
-              {/* Background pattern for visual interest */}
-              <div
-                style={{
-                  position: 'absolute',
-                  top: 0,
-                  left: 0,
-                  right: 0,
-                  bottom: 0,
-                  background:
-                    'radial-gradient(circle at top right, rgba(255,255,255,0.2) 0%, transparent 50%)',
-                  pointerEvents: 'none',
-                }}
-              />
-
-              {/* Icon container with pulse or spinner animation */}
-              <div
-                style={{
-                  width: '120px',
-                  height: '120px',
-                  backgroundColor: 'rgba(255, 255, 255, 0.3)',
-                  borderRadius: '50%',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  margin: '0 auto 32px',
-                  position: 'relative',
-                  zIndex: 2,
-                  animation:
-                    activePendingScan.phase === 'detected'
-                      ? 'pulse-scale 1s ease-in-out infinite'
-                      : 'none',
-                }}
-              >
-                {activePendingScan.phase === 'detected' ? (
-                  // Pulsing RFID signal icon for "detected" phase
-                  <svg
-                    width="80"
-                    height="80"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="white"
-                    strokeWidth="2.5"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                  >
-                    {/* NFC/RFID icon - signal waves */}
-                    <path d="M6 8.32a7.43 7.43 0 0 1 0 7.36" />
-                    <path d="M9.46 6.21a11.76 11.76 0 0 1 0 11.58" />
-                    <path d="M12.91 4.1a15.91 15.91 0 0 1 .01 15.8" />
-                    <path d="M16.37 2a20.16 20.16 0 0 1 0 20" />
-                  </svg>
-                ) : (
-                  // Spinner for "processing" phase
-                  <svg
-                    width="80"
-                    height="80"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="white"
-                    strokeWidth="2.5"
-                    strokeLinecap="round"
-                    style={{ animation: 'spin 1s linear infinite' }}
-                  >
-                    <path d="M21 12a9 9 0 1 1-6.219-8.56" />
-                  </svg>
-                )}
-              </div>
-
-              <h2
-                style={{
-                  fontSize: '36px',
-                  fontWeight: 700,
-                  marginBottom: '16px',
-                  color: '#FFFFFF',
-                  position: 'relative',
-                  zIndex: 2,
-                }}
-              >
-                {activePendingScan.phase === 'detected'
-                  ? 'Armband erkannt...'
-                  : 'Wird verarbeitet...'}
-              </h2>
-
-              <p
-                style={{
-                  fontSize: '20px',
-                  color: 'rgba(255, 255, 255, 0.9)',
-                  position: 'relative',
-                  zIndex: 2,
-                }}
-              >
-                {activePendingScan.phase === 'detected'
-                  ? 'Daten werden geladen'
-                  : 'Bitte kurz warten...'}
-              </p>
-            </ModalBase>
-          )}
+          {/* Pending Scan Modal - shows "Armband erkannt..." feedback for slow APIs (>300ms) */}
+          <PendingScanModal />
 
           {/* Main Content - Centered */}
           <div
